+++ src/app/urls.py
@@ -1,7 +1,7 @@
 # -*- coding:utf-8 -*-
 
 from django.conf.urls import include, url
-from app.views.views import ContactView
+from app.views.contact_views import ContactView
 from app.views.test_views import TestRequest
 from django.contrib import admin
